Created component for challenge goal inputs
